+++ src/rollers/rsr.js
@@ -3,6 +3,7 @@ import {
   AbilityCheckMessage,
   AbilitySaveMessage,
   AttackMessage,
+  ConcentrationMessage,
   DamageMessage,
   DeathSaveMessage,
   SkillMessage,
@@ -19,6 +20,7 @@ import {
   AbilityCheckSource,
   AbilitySaveSource,
   AttackSource,
+  ConcentrationSource,
   CriticalSource,
   DeathSaveSource,
   SkillSource,
@@ -26,6 +28,10 @@ import {
 import { showSources } from "../settings.js";
 import { debug, getTarget } from "../util.js";
 import CoreRollerHooks from "./core.js";
+
+// disable the grants.critical.range flag since RSR can't have it's critical flag changed anyways,
+// only set by the attack roll
+const distanceFn = () => Infinity;
 
 /**
  * Setup the dnd5e.preRoll hooks for use with Ready Set Roll.
@@ -67,6 +73,16 @@ export default class ReadySetRollHooks extends CoreRollerHooks {
     }
 
     if (this._doReminder(config)) new AbilitySaveReminder(actor, abilityId).updateOptions(config);
+  }
+
+  preRollConcentration(actor, options) {
+    debug("preRollConcentration hook called");
+
+    if (this._doMessages(options)) {
+      new ConcentrationMessage(actor, options.ability).addMessage(options);
+      if (showSources) new ConcentrationSource(actor, options.ability).updateOptions(options);
+    }
+    // don't need a reminder, the system will set advantage/disadvantage
   }
 
   preRollAbilityTest(actor, config, abilityId) {
@@ -125,7 +141,7 @@ export default class ReadySetRollHooks extends CoreRollerHooks {
 
     if (this._doMessages(config)) {
       new DamageMessage(item.actor, target, item).addMessage(config);
-      if (showSources) new CriticalSource(item.actor, target, item).updateOptions(config);
+      if (showSources) new CriticalSource(item.actor, target, item, distanceFn).updateOptions(config);
     }
     // don't use CriticalReminder here, it's done in another hook
   }
@@ -135,7 +151,7 @@ export default class ReadySetRollHooks extends CoreRollerHooks {
 
     // check for critical hits but set the "isCrit" property instead of the default "critical"
     const target = getTarget();
-    new CriticalReminder(item.actor, target, item).updateOptions(config, "isCrit");
+    new CriticalReminder(item.actor, target, item, distanceFn).updateOptions(config, "isCrit");
   }
 
   _doMessages({ fastForward = false }) {